Password protection

+++ ctl/Filedel.php
@@ -1,5 +1,10 @@
 <?php
 
+if (!$ctx->manage->isAdmin()) {
+    $ctx->util->redirect('main');
+    return;
+}
+
 $fileName = $ctx->util->paramGet('name');
 
 unlink("datab/$fileName");
